Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -26,7 +26,6 @@ const storage = multer.diskStorage({
 });
 
 router.get("", (req, res, next) => {
-  // console.log(req.query);
   const pageSize = +req.query.pagesize; //change to number
   const currentPage = req.query.page;
   const postQuery = Post.find();
@@ -91,21 +90,32 @@ router.put('/:id', checkAuth, multer({storage: storage}).single("image"), (req, 
     content: req.body.content,
     imagePath: imagePath,
   })
-  console.log(post);
-  Post.updateOne({_id: req.params.id}, post).then((result) => {
-    // console.log(result)
-    res.status(200).json({
-      message: "Post updated successfully!",
-    })
+  // console.log(post);
+  Post.updateOne({_id: req.params.id, creator: req.userData.userId}, post).then((result) => {
+    if (result.modifiedCount > 0) {
+      res.status(200).json({
+        message: "Post updated successfully!",
+      })
+    } else {
+      res.status(401).json({
+        message: "Not Authorized!",
+      })
+    }
   })
 });
 
 router.delete('/:id', checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id}).then((result) => {
-    console.log(result);
+  Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then((result) => {
+    if(result.deletedCount > 0) {
+
     res.status(200).json({
       message: "Post deleted successfully!",
     });
+    } else {
+      res.status(401).json({
+        message: "Not Authorized!",
+      })
+    }
   });
 });
